refactor: modify template structure
